# docs/APIs/openapi.json
{
  "openapi": "3.1.0",
  "components": {
...
    "schemas": {
...
      "AgentExecutionMetadata": {
        "description": "Documentation DTOs shared by the representative agent controllers.\n\nThese interfaces intentionally mirror the stable request and response shapes\nexposed by the Dialogue, Web, and Knowledge agents so tsoa can generate a\nreadable OpenAPI contract for Phase 1 of the documentation system.",
        "properties": {
          "duration": {
            "type": "number",
            "format": "double",
            "description": "Time spent fulfilling the request, in milliseconds."
          },
          "retryCount": {
            "type": "number",
            "format": "double",
            "description": "Number of retry attempts applied before the final response."
          }
        },
        "required": [
          "duration",
          "retryCount"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "AgentErrorEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Indicates that the proxied agent request failed."
          },
          "error": {
            "type": "string",
            "description": "Human-readable error summary returned to the caller."
          },
          "metadata": {
            "$ref": "#/components/schemas/AgentExecutionMetadata",
            "description": "Execution metadata captured for the failed request."
          }
        },
        "required": [
          "success",
          "error",
          "metadata"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "ResearchSourceItem": {
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the source document or result."
          },
          "url": {
            "type": "string",
            "description": "Canonical URL for the source."
          },
          "snippet": {
            "type": "string",
            "description": "Excerpt used in the synthesis pipeline."
          },
          "relevance": {
            "type": "number",
            "format": "double",
            "description": "Relative relevance score assigned by the Knowledge agent."
          },
          "hostname": {
            "type": "string",
            "description": "Hostname used to help explain provenance."
          }
        },
        "required": [
          "title",
          "url",
          "snippet",
          "relevance",
          "hostname"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeResearchPayload": {
        "properties": {
          "topic": {
            "type": "string",
            "description": "Topic that the Knowledge agent researched."
          },
          "summary": {
            "type": "string",
            "description": "Synthesized answer built from the top-ranked sources."
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ResearchSourceItem"
            },
            "type": "array",
            "description": "Sources retained after de-duplication and ranking."
          },
          "keyPoints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "description": "Key points extracted from the highest-value source snippets."
          },
          "confidence": {
            "type": "number",
            "format": "double",
            "description": "Confidence score for the overall synthesis."
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp recorded when the synthesis completed."
          },
          "queriesUsed": {
            "type": "number",
            "format": "double",
            "description": "Number of distinct search queries issued for the topic."
          },
          "depth": {
            "type": "string",
            "enum": [
              "quick",
              "medium",
              "deep"
            ],
            "description": "Research depth that drove the query fan-out."
          }
        },
        "required": [
          "topic",
          "summary",
          "sources",
          "keyPoints",
          "confidence",
          "timestamp",
          "queriesUsed",
          "depth"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeResearchResponseEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Indicates that the Knowledge research request succeeded."
          },
          "data": {
            "$ref": "#/components/schemas/KnowledgeResearchPayload",
            "description": "Research payload generated by the Knowledge agent."
          },
          "metadata": {
            "$ref": "#/components/schemas/AgentExecutionMetadata",
            "description": "Execution metadata captured for the request."
          }
        },
        "required": [
          "success",
          "data",
          "metadata"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeResearchRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "description": "Topic or question that should be researched from multiple angles."
          },
          "depth": {
            "type": "string",
            "enum": [
              "quick",
              "medium",
              "deep"
            ],
            "description": "Depth of research requested for the topic."
          }
        },
        "required": [
          "topic"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeFactCheckPayload": {
        "properties": {
          "claim": {
            "type": "string",
            "description": "Claim that was evaluated."
          },
          "verdict": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "DISPUTED",
              "UNVERIFIED"
            ],
            "description": "Final verdict calculated by the Knowledge agent."
          },
          "confidence": {
            "type": "number",
            "format": "double",
            "description": "Confidence score for the verdict."
          },
          "confirmingSources": {
            "items": {
              "$ref": "#/components/schemas/ResearchSourceItem"
            },
            "type": "array",
            "description": "Sources that support the claim."
          },
          "contradictingSources": {
            "items": {
              "$ref": "#/components/schemas/ResearchSourceItem"
            },
            "type": "array",
            "description": "Sources that dispute the claim."
          },
          "neutralSources": {
            "items": {
              "$ref": "#/components/schemas/ResearchSourceItem"
            },
            "type": "array",
            "description": "Sources that were relevant but inconclusive."
          },
          "explanation": {
            "type": "string",
            "description": "Narrative explanation of the verdict."
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp recorded when the fact-check completed."
          }
        },
        "required": [
          "claim",
          "verdict",
          "confidence",
          "confirmingSources",
          "contradictingSources",
          "neutralSources",
          "explanation",
          "timestamp"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeFactCheckResponseEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Indicates that the fact-check request succeeded."
          },
          "data": {
            "$ref": "#/components/schemas/KnowledgeFactCheckPayload",
            "description": "Fact-check result returned by the Knowledge agent."
          },
          "metadata": {
            "$ref": "#/components/schemas/AgentExecutionMetadata",
            "description": "Execution metadata captured for the request."
          }
        },
        "required": [
          "success",
          "data",
          "metadata"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeFactCheckRequest": {
        "properties": {
          "claim": {
            "type": "string",
            "description": "Claim that should be verified across multiple sources."
          },
          "sources": {
            "type": "number",
            "format": "double",
            "description": "Number of sources to inspect per fact-check query."
          }
        },
        "required": [
          "claim"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "SummarySourceItem": {
        "properties": {
          "title": {
            "type": "string",
            "description": "Human-readable source title."
          },
          "url": {
            "type": "string",
            "description": "Canonical URL for the source."
          }
        },
        "required": [
          "title",
          "url"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeSummaryPayload": {
        "properties": {
          "query": {
            "type": "string",
            "description": "Query that was summarized."
          },
          "summary": {
            "type": "string",
            "description": "Combined synthesis built from the fetched sources."
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SummarySourceItem"
            },
            "type": "array",
            "description": "Sources cited by the summarized response."
          },
          "wordCount": {
            "type": "number",
            "format": "double",
            "description": "Word count of the final summary text."
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp recorded when summarization completed."
          }
        },
        "required": [
          "query",
          "summary",
          "sources",
          "wordCount",
          "timestamp"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeSummaryResponseEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Indicates that the summarization request succeeded."
          },
          "data": {
            "$ref": "#/components/schemas/KnowledgeSummaryPayload",
            "description": "Summarization payload returned by the Knowledge agent."
          },
          "metadata": {
            "$ref": "#/components/schemas/AgentExecutionMetadata",
            "description": "Execution metadata captured for the request."
          }
        },
        "required": [
          "success",
          "data",
          "metadata"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "KnowledgeSummarizeRequest": {
        "properties": {
          "query": {
            "type": "string",
            "description": "Topic or search query to summarize from the web."
          },
          "maxSources": {
            "type": "number",
            "format": "double",
            "description": "Maximum number of source documents to incorporate."
          }
        },
        "required": [
          "query"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Jarvis Representative Agent API",
    "version": "4.0.0",
    "description": "Phase 1 OpenAPI surface generated from representative Dialogue, Web, and Knowledge controllers for the Jarvis documentation system foundation.",
    "contact": {}
  },
  "paths": {
...
    "/api/knowledge/research": {
      "post": {
        "operationId": "researchWithKnowledgeAgent",
        "responses": {
          "200": {
            "description": "Research result returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeResearchResponseEnvelope"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "success": true,
                      "data": {
                        "topic": "documentation drift prevention",
                        "summary": "Documentation drift is reduced by generating docs from source code.",
                        "sources": [
                          {
                            "title": "Generated documentation guide",
                            "url": "https://example.com/generated-docs",
                            "snippet": "A representative source snippet.",
                            "relevance": 0.93,
                            "hostname": "example.com"
                          }
                        ],
                        "keyPoints": [
                          "Generate API reference from code.",
                          "Keep architecture rationale in ADRs."
                        ],
                        "confidence": 0.86,
                        "timestamp": "2026-03-22T11:05:24.498Z",
                        "queriesUsed": 3,
                        "depth": "medium"
                      },
                      "metadata": {
                        "duration": 618,
                        "retryCount": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request body is missing a topic",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "The Knowledge agent research request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorEnvelope"
                }
              }
            }
          }
        },
        "description": "Research a topic from multiple angles and synthesize a citation-friendly\nanswer from the highest-value retrieved sources.",
        "tags": [
          "Knowledge"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeResearchRequest"
              }
            }
          }
        }
      }
    },
    "/api/knowledge/fact-check": {
      "post": {
        "operationId": "factCheckWithKnowledgeAgent",
        "responses": {
          "200": {
            "description": "Fact-check result returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeFactCheckResponseEnvelope"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "success": true,
                      "data": {
                        "claim": "Generated documentation reduces drift.",
                        "verdict": "CONFIRMED",
                        "confidence": 0.82,
                        "confirmingSources": [
                          {
                            "title": "Documentation best practices",
                            "url": "https://example.com/best-practices",
                            "snippet": "Generated documentation reduces stale reference material.",
                            "relevance": 0.88,
                            "hostname": "example.com"
                          }
                        ],
                        "contradictingSources": [],
                        "neutralSources": [],
                        "explanation": "Most retrieved sources support the claim with minimal contradiction.",
                        "timestamp": "2026-03-22T11:05:24.498Z"
                      },
                      "metadata": {
                        "duration": 744,
                        "retryCount": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request body is missing a claim",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "The Knowledge agent fact-check request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorEnvelope"
                }
              }
            }
          }
        },
        "description": "Verify a claim across multiple sources and return a confidence-weighted\nverdict describing whether the claim is supported, disputed, or unresolved.",
        "tags": [
          "Knowledge"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeFactCheckRequest"
              }
            }
          }
        }
      }
    },
    "/api/knowledge/summarize": {
      "post": {
        "operationId": "summarizeWithKnowledgeAgent",
        "responses": {
          "200": {
            "description": "Summary returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSummaryResponseEnvelope"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "success": true,
                      "data": {
                        "query": "Jarvis documentation system phase 1",
                        "summary": "Phase 1 establishes TypeDoc, representative OpenAPI, ADRs, and a glossary.",
                        "sources": [
                          {
                            "title": "Implementation plan",
                            "url": "https://example.com/implementation-plan"
                          }
                        ],
                        "wordCount": 12,
                        "timestamp": "2026-03-22T11:05:24.498Z"
                      },
                      "metadata": {
                        "duration": 403,
                        "retryCount": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request body is missing a query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "The Knowledge agent summarize request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorEnvelope"
                }
              }
            }
          }
        },
        "description": "Summarize current web information about a topic into a concise digest with\nsource references and a computed word count.",
        "tags": [
          "Knowledge"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSummarizeRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}